postgres_to_es/requirements.txt
certifi==2020.12.5
elasticsearch==7.11.0
psycopg2-binary==2.8.6
redis==3.5.3
urllib3==1.26.3
